requirements.txt
mcp>=0.5.0
google-genai>=0.3.0
pillow>=10.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
pydantic>=2.0.0
